# apis/v1/server/openapi.gen.json
{
  "components": {
    "parameters": {
      "Limit": {
        "in": "query",
        "name": "limit",
        "schema": {
          "default": 100,
          "maximum": 1000,
          "minimum": 1,
          "type": "integer"
        }
      },
      "Offset": {
        "in": "query",
        "name": "offset",
        "schema": {
          "default": 0,
          "minimum": 0,
          "type": "integer"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/apis_v1_components_ErrorResponse"
            }
          }
        },
        "description": "Invalid request."
      },
...
      "InternalError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/apis_v1_components_ErrorResponse"
            }
          }
        },
        "description": "Internal request processing failed."
      },
...
      "PermissionDenied": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/apis_v1_components_ErrorResponse"
            }
          }
        },
        "description": "Permission denied."
      },
...
      "Unauthenticated": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/apis_v1_components_ErrorResponse"
            }
          }
        },
        "description": "Authentication failed.",
        "headers": {
          "WWW-Authenticate": {
            "required": true,
            "schema": {
              "example": "Bearer",
              "type": "string"
            }
          }
        }
      },
...
    },
    "schemas": {
      "CreateProfilingJobRequest": {
        "additionalProperties": false,
        "properties": {
          "binary_match_path": {
            "minLength": 1,
            "type": "string"
          },
          "container_id": {
            "minLength": 1,
            "type": "string"
          },
          "duration_seconds": {
            "format": "int64",
            "minimum": 1,
            "type": "integer"
          },
          "hostname": {
            "minLength": 1,
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/ProfilingLanguage"
          },
          "mode": {
            "$ref": "#/components/schemas/ProfilingMode"
          },
          "scope": {
            "$ref": "#/components/schemas/apis_v1_components_ObservationScope"
          },
          "type": {
            "$ref": "#/components/schemas/ProfilingType"
          }
        },
        "required": [
          "hostname",
          "duration_seconds",
          "scope",
          "type",
          "language",
          "mode"
        ],
        "type": "object"
      },
...
      "ProfilingJobListResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": false,
            "properties": {
              "items": {
                "items": {
                  "$ref": "#/components/schemas/ProfilingJob"
                },
                "type": "array"
              },
              "limit": {
                "type": "integer"
              },
              "offset": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              }
            },
            "required": [
              "items",
              "total",
              "limit",
              "offset"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "$ref": "#/components/securitySchemes/apis_v1_components_BearerAuth"
      },
      "apis_v1_components_BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Huatuo Apiserver API",
    "version": "v1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/profiling": {
      "get": {
        "operationId": "listProfilingJobs",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilingJobListResponse"
                }
              }
            },
            "description": "Profiling Jobs."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "profiling"
        ]
      },
      "post": {
        "operationId": "createProfilingJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfilingJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilingJobResponse"
                }
              }
            },
            "description": "Profiling Job created.",
            "headers": {
              "Location": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "profiling"
        ]
      }
    },
    "/v1/profiling/capabilities": {
      "get": {
        "operationId": "getProfilingCapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilingCapabilitiesResponse"
                }
              }
            },
            "description": "Static profiling capabilities."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "profiling"
        ]
      }
    },
    "/v1/profiling/{request_id}": {
      "get": {
        "operationId": "getProfilingJob",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilingJobResponse"
                }
              }
            },
            "description": "Profiling Job."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "profiling"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ]
    },
    "/v1/profiling/{request_id}/raw": {
      "get": {
        "operationId": "getRawProfiles",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawProfilePageResponse"
                }
              }
            },
            "description": "Raw profile page."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "profiling"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ]
    },
    "/v1/profiling/{request_id}/stop": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ],
      "post": {
        "operationId": "stopProfilingJob",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilingJobResponse"
                }
              }
            },
            "description": "Current Profiling Job after applying the stop intent."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "profiling"
        ]
      }
    },
    "/v1/tracing": {
      "get": {
        "operationId": "listTracingJobs",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracingJobListResponse"
                }
              }
            },
            "description": "Tracing Jobs."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "tracing"
        ]
      },
      "post": {
        "operationId": "createTracingJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTracingJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracingJobResponse"
                }
              }
            },
            "description": "Tracing Job created.",
            "headers": {
              "Location": {
                "required": true,
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "413": {
            "$ref": "#/components/responses/RequestTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "tracing"
        ]
      }
    },
    "/v1/tracing/capabilities": {
      "get": {
        "operationId": "getTracingCapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracingCapabilitiesResponse"
                }
              }
            },
            "description": "Static tracing capabilities."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "tracing"
        ]
      }
    },
    "/v1/tracing/{request_id}": {
      "get": {
        "operationId": "getTracingJob",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracingJobResponse"
                }
              }
            },
            "description": "Tracing Job."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "tracing"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ]
    },
    "/v1/tracing/{request_id}/stop": {
      "parameters": [
        {
          "$ref": "#/components/parameters/RequestID"
        }
      ],
      "post": {
        "operationId": "stopTracingJob",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracingJobResponse"
                }
              }
            },
            "description": "Current Tracing Job after applying the stop intent."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/PermissionDenied"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "tracing"
        ]
      }
    }
  },
...
}